Adding new option with stellar-core command : --bitcoind
