+++ stellar-core.cfg
@@ -14,6 +14,9 @@ NODE_IS_VALIDATOR=true
 #DATABASE="sqlite3://stellar.db"
 DATABASE="postgresql://dbname=stellar_node01_db host=localhost user=stellar_node_user password=testdb"
 
+#Bitcoin realted configs
+BITCOIN_REGTEST=true
+BITCOIN_DATADIR="/home/vishswasb/zagg/node1/"
 
 #Bucket directory path
 BUCKET_DIR_PATH = "buckets"
@@ -34,3 +37,4 @@ VALIDATORS=["$zaggnode01"]
 get="cp /tmp/stellar-core/history/vs/{0} {1}"
 put="cp {0} /tmp/stellar-core/history/vs/{1}"
 mkdir="mkdir -p /tmp/stellar-core/history/vs/{0}"
+
